% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.r
\name{chain.data}
\alias{chain.data}
\title{Get a dataset from a chain}
\usage{
chain.data(chain)
}
\arguments{
\item{chain}{The chain to build a dataset from}
}
\value{
A dataset comprised of all of the chain's
members' data put together. The order of the rows
in the dataset is not specified.
}
\description{
Take the data for each member of the given chain
and assemble it together in a dataset.
}
