% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{get.fn}
\alias{get.fn}
\title{grab a function based on its name}
\usage{
get.fn(fn, env = parent.frame())
}
\arguments{
\item{fn}{the function to search for}

\item{env}{the environment to start searching in}
}
\value{
fn, if fn is already a function; otherwise, the first function found
        in env or one of its parents whose name is fn
}
\description{
helper to grab a fn that is passed in as an argument
}
\details{
this is based on Hadley Wickham's response to an SO
post: \url{http://stackoverflow.com/questions/14183766/match-fun-provide-error-with-functions-defined-inside-functions}
with some minor modifications
}
\keyword{internal}

