% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaled_bootstrap.r
\name{get.rescaled.bootstrap.weights}
\alias{get.rescaled.bootstrap.weights}
\title{rescaled.bootstrap.weights}
\usage{
get.rescaled.bootstrap.weights(
  survey.data,
  survey.design,
  idvar = NULL,
  weights = NULL,
  parallel = FALSE,
  paropts = NULL,
  num.reps = 1,
  include_scaling_factors = FALSE,
  include_cc = FALSE
)
}
\arguments{
\item{survey.data}{The dataset to use}

\item{survey.design}{A formula describing the design of the survey (see Details)}

\item{idvar}{the name of the column in \code{survey.data} that has the respondent id}

\item{weights}{Survey weights to be rescaled (or \code{NULL}, if none)}

\item{parallel}{If \code{TRUE}, use parallelization (via \code{plyr})}

\item{paropts}{An optional list of arguments passed on to \code{plyr} to control
details of parallelization}

\item{num.reps}{The number of bootstrap replication samples to draw}

\item{include_scaling_factors}{If \code{TRUE}, include \code{weight_scaling_factor} in the output}

\item{include_cc}{If \code{TRUE}, include cluster counts in the output. These cluster counts
are needed for jackknife after bootstrap calculations}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A list with entries:

* `orig_weights` - A dataframe with rows corresponding to the original
   `survey.data` and a column with the original (un rescaled) weight for the obs.
   Can be useful for debugging. 

* `boot_weights` - A dataframe with rows corresponding to the original
   `survey.data` and one column for each bootstrap rep. The entries have
   the rescaled bootstrap weights for each row and bootstrap rep. 

* `weight_scaling_factor` - (if `include_scaling_factors` is `TRUE`). A dataframe with rows corresponding to the original
   `survey.data` and one column for each bootstrap rep. The entries have
   the values by which the original weights were scaled to produce `boot_weights`.
   You typically won't need to use this dataframe, but it can be helpful for debugging.

* `cluster_counts` - (if `include_cc` is `TRUE`). A list with `num.reps` entries.
  Each entry is a dataset which has the variable '.cluster_id',
  the values of columns that specify the PSU (from the `survey.design` formula),
  and `cluster_count` (the number of times the given PSU was resampled)) 
}\if{html}{\out{</div>}}
}
\description{
Get a dataset of rescaled bootstrap weights.
}
\details{
This is a new function being added as part of a refactor
in v.0.2. This function focuses on being able to return
a dataframe with bootstrap weights, and optionally also
a dataframe with cluster/PSU inclusion counts.

\code{survey.design} is a formula of the form

\code{weight ~ psu_vars + strata(strata_vars)}

where:
\itemize{
\item \code{weight} is the variable with the survey weights
\item \code{psu_vars} has the form \code{psu_v1 + psu_v2 + ...}, where primary
sampling units (PSUs) are determined by \code{psu_v1}, etc
\item \code{strata_vars} has the form \code{strata_v1 + strata_v2 + ...}, which
determine strata
}

Note that we assume that the formula uniquely specifies PSUs.
This will always be true if the PSUs were selected without replacement.
If they were selected with replacement, then it will be necessary
to make each realization of a given PSU in the sample a unique id.
The code below assumes that all observations within
each PSU (as identified by the design formula) are from the same draw
of the PSU.

The rescaled bootstrap technique works by adjusting the
estimation weights based on the number of times each
row is included in the resamples. If a row is never selected,
it is still included in the returned results, but its weight
will be set to 0. It is therefore important to use estimators
that make use of the estimation weights on the resampled
datasets.

We always take m_i = n_i - 1, according to the advice presented
in Rao and Wu (1988) and Rust and Rao (1996).

(This is a C++ version; a previous version, written in pure R,
is called \code{\link[=rescaled.bootstrap.sample.pureR]{rescaled.bootstrap.sample.pureR()}} )

References:
\itemize{
\item Rust, Keith F., and J. N. K. Rao. "Variance estimation for complex surveys
using replication techniques." \emph{Statistical methods in medical research}
5.3 (1996): 283-310.
\item Rao, Jon NK, and C. F. J. Wu. "Resampling inference with complex survey
data." \emph{Journal of the American Statistical Association}
83.401 (1988): 231-241.
}
}
\examples{

survey <- MU284.complex.surveys[[1]]
boot_surveys <- get.rescaled.bootstrap.weights(survey.data = survey,
                                               survey.design = ~ CL,
                                               num.reps = 2)

}
