% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{get.var}
\alias{get.var}
\title{get a variable from a dataframe or vector}
\usage{
get.var(survey.data, var, default = NA)

get.var(survey.data, var, default = NA)
}
\arguments{
\item{survey.data}{the survey dataset}

\item{var}{either NULL, a column name, or a vector of values}

\item{default}{the default value to fill in if the variable
is not found}

\item{survey.data}{the survey dataset}

\item{var}{either NULL, a column name, or a vector of values}

\item{default}{the default value to fill in if the variable
is not found}
}
\value{
a vector of values whose length is the same as the
        number of rows in survey.data; if var is NULL, this has
        the default values

a vector of values whose length is the same as the
        number of rows in survey.data; if var is NULL, this has
        the default values
}
\description{
this function was written because a few of the estimator functions
need to use weights, and there are several cases to handle:
the user could pass in a column name, a vector of weights, or
nothing (in which case, the weights should default to 1 for each
row in the dataset). for the special case of getting weights, look
at the curried fn get.weights (right below)

this function was written because a few of the estimator functions
need to use weights, and there are several cases to handle:
the user could pass in a column name, a vector of weights, or
nothing (in which case, the weights should default to 1 for each
row in the dataset). for the special case of getting weights, look
at the curried fn get.weights (right below)
}
\keyword{internal}

