% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.r
\name{max.depth}
\alias{max.depth}
\title{Get the height (maximum depth) of a chain}
\usage{
\method{max}{depth}(chain)
}
\arguments{
\item{chain}{The chain object}
}
\value{
The maximum depth of the chain
}
\description{
Get the height (maximum depth) of a chain
}
