% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{MU284.boot.res.summ}
\alias{MU284.boot.res.summ}
\title{Benchmarks for unit tests}
\format{
A list with 10 data frames, each with 15 rows and 11 columns:
\describe{
\item{mean.TS82.hat, ..., sd.R.RMT85.P85.hat}{summaries for each estimand}
}
}
\description{
Benchmark results to use in unit tests; these are based on
\link{MU284.complex.surveys}.
}
