% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{MU284.estimator.fn}
\alias{MU284.estimator.fn}
\title{MU284.estimator.fn}
\usage{
MU284.estimator.fn(survey.data, weights)
}
\arguments{
\item{survey.data}{the survey dataset}

\item{weights}{a vector with the survey weights}
}
\value{
a data.frame with one row and two columns:
\itemize{
\item \code{TS82.hat} - the estimated total of \code{S82}
\item \code{R.RMT85.P85.hat} - the estimated ratio of \code{RMT85} / \code{P85}
}
}
\description{
Produce estimates from a simulated sample survey of the \link{MU284} population.
Used in package tests and examples.
}
