% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{MU284.estimator.summary.fn}
\alias{MU284.estimator.summary.fn}
\title{MU284.estimator.summary.fn}
\usage{
MU284.estimator.summary.fn(res)
}
\arguments{
\item{res}{a dataframe whose rows are the results of calling \code{\link[=MU284.estimator.fn]{MU284.estimator.fn()}}}
}
\value{
the same dataframe
}
\description{
Summarize results from \code{\link[=MU284.estimator.fn]{MU284.estimator.fn()}} applied to many surveys.
(This is a dummy function, used for tests)
}
