% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{MU284}
\alias{MU284}
\title{The MU284 Population dataset}
\format{
A data frame with 284 rows and 11 columns:
\describe{
\item{LABEL}{ID}
\item{P85}{Population in 1985}
\item{P75}{Population in 1975}
\item{RMT85}{Municipal tax revenue in 1985}
\item{CS82}{Number of Conservative seats in municipal council}
\item{SS82}{Number of Social-Democratic seats in municipal council}
\item{S82}{Total number of seats in municipal council}
\item{ME84}{Number of municipal employees}
\item{REV84}{Real estate values according to 1984 assessment}
\item{REG}{Geographic location indicator}
\item{CL}{Cluster indicator (neighboring municipalities are clustered together)}
}
}
\source{
'Model Assisted Survey Sampling' by Sarndal, Swensson, and Wretman (2003, ISBN:0387406204)
}
\description{
A dataset containing information about Sweden's 284 municipalities. This
dataset comes from Model-Assisted Survey Sampling by
Sarndal, Swensson, and Wretman (2003, ISBN:0387406204). The columns are:
}
