% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{MU284.surveys}
\alias{MU284.surveys}
\title{Simulated sample surveys drawn from the MU284 Population}
\format{
A data frame with 284 rows and 11 columns:
\describe{
\item{LABEL, ..., CL}{Same as MU284 dataset}
\item{sample_weight}{The sampling weight for the row}
}
}
\description{
A list with 15 sample surveys drawn from the \link{MU284} dataset.
}
