% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{parse_design}
\alias{parse_design}
\title{Parse a formula that describes the design of a survey}
\usage{
parse_design(formula)
}
\arguments{
\item{formula}{a formula describing the sample design (see Description of \code{\link[=bootstrap.estimates]{bootstrap.estimates()}})}
}
\value{
a list with entries \code{psu.formula} and \code{strata.formula}
}
\description{
Parse a formula of the form
\code{~ psu_v1 + psu_v2 + ... + strata(strata_v1 + strata_v2 + ...)}
into a PSU formula and a strata formula.
}
\keyword{internal}
