% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_bootstrap.r
\name{rds.mc.boot.draws}
\alias{rds.mc.boot.draws}
\title{draw RDS bootstrap resamples using the
algorithm in Salganik 2006 (TODO PROPER CITE)}
\usage{
rds.mc.boot.draws(chains, mm, dd, num.reps)
}
\arguments{
\item{chains}{a list with the chains constructed from the survey
using \code{make.chain}}

\item{mm}{the mixing model}

\item{dd}{the degree distributions}

\item{num.reps}{the number of bootstrap resamples we want}
}
\value{
a list of length \code{num.reps}; each entry in
the list has one bootstrap-resampled dataset
}
\description{
this algorithm picks a respondent from the survey
to be a seed uniformly at random. it then generates
a bootstrap draw by simulating the markov process
forward for n steps, where n is the size of the draw
required.
}
\details{
if you wish the bootstrap dataset to end up with
variables from the original dataset other than the
traits and degree, then you must specify this when
you construct \code{dd} using the 
'\code{estimate.degree.distns} function.

TODO be sure to comment the broken-out trait variables
     (ie these could all be different from the originals)
}

