% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_weights.r
\name{rescaled.bootstrap.weights}
\alias{rescaled.bootstrap.weights}
\title{rescaled.bootstrap.weights}
\usage{
rescaled.bootstrap.weights(
  survey.data,
  survey.design,
  num.reps,
  weights = NULL,
  idvar,
  verbose = TRUE,
  parallel = FALSE,
  paropts = NULL
)
}
\arguments{
\item{survey.data}{the dataset to use}

\item{survey.design}{a formula describing the design of the survey
(see below - TODO)}

\item{num.reps}{the number of bootstrap replication samples to draw}

\item{weights}{weights to use in estimation (or NULL, if none)}

\item{idvar}{the name of the column in \code{survey.data} that has the respondent id}

\item{verbose}{if TRUE, produce lots of feedback about what is going on}

\item{parallel}{if TRUE, use the plyr library's .parallel argument to
produce bootstrap resamples and estimates in parallel}

\item{paropts}{if not NULL, additional arguments to pass along to the
parallelization routine}
}
\value{
if no summary.fn is specified, then return the list of estimates
        produced by estimator.fn; if summary.fn is specified, then return
        its output
}
\description{
This function creates a dataset with rescaled bootstrap weights;
it can be a helpful alternative to \code{bootstrap.estimates} in some situations
}
\examples{
\donttest{
# code goes here

}
}
