% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveybootstrap_package.r
\docType{package}
\name{surveybootstrap}
\alias{surveybootstrap-package}
\alias{surveybootstrap}
\alias{package-surveybootstrap}
\title{Survey bootstrap variance estimators}
\description{
\code{surveybootstrap} has methods for analyzing data that were collected
using network reporting techniques. It includes estimators appropriate for
the simple bootstrap and the rescaled bootstrap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dfeehan.github.io/surveybootstrap/}
  \item \url{https://github.com/dfeehan/surveybootstrap}
  \item \url{http://dennisfeehan.org/surveybootstrap/}
  \item Report bugs at \url{https://github.com/dfeehan/surveybootstrap/issues}
}

}
\author{
\strong{Maintainer}: Dennis M. Feehan \email{feehan@berkeley.edu}

Other contributors:
\itemize{
  \item Matthew J. Salganik \email{mjs3@princeton.edu} [thesis advisor]
}

}
