% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.r
\name{traits.to.string}
\alias{traits.to.string}
\title{take a set of traits and turn into a string}
\usage{
traits.to.string(data, traits, na.action = "drop", sep = ".")
}
\arguments{
\item{data}{the respondent info}

\item{traits}{the names of the traits to build the model on}

\item{na.action}{for now, defaults to 'drop' (meaning all rows of data
with any missingness on the traits are dropped). anything else
NAs are treated like any other value}

\item{sep}{the separator character used to combine values}
}
\value{
a list whose entries are \code{used.idx}, which indicates
which rows from the original dataset were used (may not be all of
if there is missingness); and \code{traits}, which has the string
version of the traits
}
\description{
this is a helper function that is useful when we wish
to make several traits into one variable
}
\keyword{internal}

