% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{vcat}
\alias{vcat}
\title{Only prints things out in verbose mode}
\usage{
vcat(verbose = TRUE, ...)
}
\arguments{
\item{verbose}{If \code{TRUE}, print things out; otherwise, do nothing}

\item{...}{Arguments to pass to cat if verbose is \code{TRUE}}
}
\description{
Only prints things out in verbose mode
}
\keyword{internal}
