% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.r
\name{chain.size}
\alias{chain.size}
\title{Get the size of a chain}
\usage{
chain.size(chain)
}
\arguments{
\item{chain}{The chain object}
}
\value{
The number of respondents involved in
the chain
}
\description{
Count the total number of respondents in the chain
and return it
}
