% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{get.fn}
\alias{get.fn}
\title{Grab a function based on its name}
\usage{
get.fn(fn, env = parent.frame())
}
\arguments{
\item{fn}{The function to search for}

\item{env}{The environment to start searching in}
}
\value{
\code{fn}, if \code{fn} is already a function; otherwise, the first function found
in env or one of its parents whose name is \code{fn}
}
\description{
Helper to grab a function that is passed in as an argument
}
\details{
This is based on Hadley Wickham's response to an SO
post:
\url{https://stackoverflow.com/questions/14183766/match-fun-provide-error-with-functions-defined-inside-functions}
with some minor modifications
}
\keyword{internal}
