% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{get.weights}
\alias{get.weights}
\title{Get the weights column from a dataframe}
\usage{
get.weights(...)
}
\arguments{
\item{...}{(this is a function curried from \code{\link[=get.var]{get.var()}})}
}
\description{
This is the same as \code{\link[=get.var]{get.var()}} with the default value set to 1
instead of \code{NA}
}
\keyword{internal}
