% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.r
\name{is.child.ct}
\alias{is.child.ct}
\title{determine whether or not one id is a parent of another}
\usage{
is.child.ct(id, seed.id)
}
\arguments{
\item{id}{the id of the potential child}

\item{seed.id}{the id of the potential parent}
}
\value{
TRUE if \code{id} is the direct descendant of \code{seed.id}
and FALSE otherwise
}
\description{
this function allows us to determine which ids are
directly descended from which other ones. it is the only part
of the code that relies on the ID format used by the
Curitiba study (TODO CITE); by modifying this function,
it shold be possible to adapt this code to another study
}

