% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.r
\name{make.chain}
\alias{make.chain}
\title{build an RDS seed's chain from the dataset}
\usage{
make.chain(seed.id, survey.data, is.child.fn = is.child.ct)
}
\arguments{
\item{seed.id}{the id of the seed whose chain we
wish to build from the dataset}

\item{survey.data}{the dataset}

\item{is.child.fn}{a function which takes two ids as arguments;
it is expected to return TRUE if the second argument is the parent of the
first, and FALSE otherwise. it defaults to \code{\link{is.child.ct}}}
}
\value{
info
}
\description{
text
TODO assumes that the chain is a tree (no loops)
}

