% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.r
\name{max.depth}
\alias{max.depth}
\title{get the height (maximum depth) of a chain}
\usage{
\method{max}{depth}(chain)
}
\arguments{
\item{chain}{the chain object}
}
\value{
the maximum depth of the chain
}
\description{
get the height (maximum depth) of a chain
}

