% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds.r
\name{mc.sim}
\alias{mc.sim}
\title{Run a markov model}
\usage{
mc.sim(mm, start, n)
}
\arguments{
\item{mm}{The markov model object returned by \code{\link[=estimate.mixing]{estimate.mixing()}}}

\item{start}{The name of the state to start in}

\item{n}{The number of time-steps to run through}
}
\value{
A vector with the state visited at each time step. the first entry
has the starting state
}
\description{
Run a given markov model for n time steps, starting
at a specified state.
}
\details{
This uses the markov model produced by \code{\link[=estimate.mixing]{estimate.mixing()}}
}
