% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveybootstrap_package.r
\docType{data}
\name{MU284}
\alias{MU284}
\title{MU284 population}
\description{
Data used in unit tests for variance estimation.
See TODO-Sarndal TODO-sampling package
TODO-doc describing unit tests
}

