% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{MU284.surveys}
\alias{MU284.surveys}
\title{Simulated sample surveys drawn from the MU284 Population}
\format{
A list with 10 data frames, each with 15 rows and 11 columns:
\describe{
\item{LABEL, ..., CL}{Same as MU284 dataset}
\item{sample_weight}{The sampling weight for the row}
}
}
\description{
A list with 10 sample surveys with sample size 15 drawn from the \link{MU284}
dataset using simple random sampling with replacment.
}
