% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveybootstrap_package.r
\docType{package}
\name{surveybootstrap}
\alias{surveybootstrap}
\alias{package-surveybootstrap}
\title{Survey bootstrap variance estimators}
\description{
\code{surveybootstrap} has methods for analyzing data that were collected
using network reporting techniques. It includes estimators appropriate for
the simple bootstrap and the rescaled bootstrap.
}
