% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{vcat}
\alias{vcat}
\title{only prints things out in verbose mode}
\usage{
vcat(verbose = TRUE, ...)
}
\arguments{
\item{verbose}{if TRUE, print things out; otherwise, do nothing}

\item{...}{arguments to pass to cat if verbose is TRUE}
}
\description{
only prints things out in verbose mode
}
\keyword{internal}

