% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.r
\name{weighted.mean}
\alias{weighted.mean}
\title{Compute the weighted mean}
\usage{
weighted.mean(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{The vector of values}

\item{w}{The vector of weights}

\item{na.rm}{if \code{TRUE}, only consider elements of \code{x} that are not missing
(and their corresponding entries in \code{w}). Defaults to \code{FALSE}.}
}
\value{
The weighted mean
}
\description{
Given a vector of values and a vector of weights, compute the
weighted mean
}
\keyword{internal}
